wip towards sibilant and pseudops split  [ci skip]
